orchestrator/pyproject.toml
[project]
name = "orchestrator"
version = "0.1.0"
description = "Generic skill orchestration library with LangGraph and LlamaIndex"
requires-python = ">=3.14"
dependencies = [
  "langgraph>=0.4.0",
  "llama-index>=0.12.0",
  "llama-index-core>=0.12.0",
  "llama-index-embeddings-openai>=0.3.0",
  "llama-index-vector-stores-faiss>=0.3.0",
  "openai>=1.82.0",
  "pydantic>=2.11.0",
  "pyyaml>=6.0",
  "faiss-cpu>=1.11.0",
  "numpy>=2.2.0",
  "sentence-transformers>=4.1.0",
]

[project.optional-dependencies]
qdrant = [
  "qdrant-client>=1.14.0",
  "llama-index-vector-stores-qdrant>=0.4.0",
]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.26.0",
  "pytest-cov>=6.0",
  "ruff>=0.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/orchestrator"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=orchestrator --cov-report=term-missing --cov-fail-under=80"

[tool.coverage.run]
source = ["orchestrator"]

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "if TYPE_CHECKING:",
  "if __name__",
]
